# Conflicts:
#	app/src/main/AndroidManifest.xml

+++ app/src/main/java/com/example/chillinapp/data/stress/FirebaseStressDataDao.kt
@@ -29,7 +29,6 @@ class FirebaseStressDataDao {
         val email= user?.email
         val userDocument= email?.let { accountCollection.document(it) }
 
-
         return try {
             var i:Int=0
             // Insert raw data for each sample
@@ -40,11 +39,10 @@ class FirebaseStressDataDao {
                     "heartrateSensor" to data.heartrateSensor,
                     "skinTemperatureSensor" to data.skinTemperatureSensor,
                 )
+
                 Log.d("Insert", "Insert completed")
                 rawDocument?.set(rawData)?.await()
                 i++
-
-
             }
 
             val fastreturn=fastInsert(stressData)
@@ -54,9 +52,9 @@ class FirebaseStressDataDao {
                 fastreturn
         } catch (e:Exception){
             ServiceResult(false,null,StressErrorType.NETWORK_ERROR)
-
         }
     }
+
     suspend fun getRawData(n: Int): ServiceResult <List <StressRawData>,StressErrorType> {
         val user = auth.currentUser
         val email = user?.email
@@ -72,27 +70,26 @@ class FirebaseStressDataDao {
             querySnapshot?.forEach { document ->
                 val timestamp = document.id.toLongOrNull()
                 if (timestamp != null) {
-                    val heartrateSensor:Double = (document.get("heartrateSensor") as Double)
+                    val heartrateSensor:Float = (document.get("heartrateSensor") as Float)
                     val skinTemperatureSensor : Double= (document.get("skinTemperatureSensor" ) as Double)
 
                     // Costruisci l'oggetto StressRawData e aggiungilo alla lista
                     val stressRawData = StressRawData(timestamp, heartrateSensor, skinTemperatureSensor)
                     rawDataList.add(stressRawData)
                 }
-
             }
 
             ServiceResult(true, rawDataList, null)
         } catch (e: Exception) {
             ServiceResult(false, null, StressErrorType.NETWORK_ERROR)
-
-
         }
     }
+
     suspend fun fastInsert(stressData: List<StressRawData>) :ServiceResult<Unit,StressErrorType>{
         val email= auth.currentUser?.email
         val key= email?.substringBefore("@")
         val rawdatareference= key?.let { dbreference.child(it).child("RawData") }
+
         if (rawdatareference != null) {
             return try {
                 rawdatareference.removeValue().await()
@@ -107,12 +104,9 @@ class FirebaseStressDataDao {
             }
         }
        else
-
-        return ServiceResult(success = false, data = null, error = StressErrorType.NOACCOUNT)
-
-
-
+           return ServiceResult(success = false, data = null, error = StressErrorType.NOACCOUNT)
     }
+
     suspend fun fastget():ServiceResult<List<StressRawData>,StressErrorType>{
         val user=auth.currentUser
         val email=user?.email
@@ -123,21 +117,18 @@ class FirebaseStressDataDao {
             val stressDataList= mutableListOf<StressRawData>()
             snapshot?.children?.forEach{ childSnapshot->
                 val timestamp=childSnapshot.key?.toLongOrNull()
-                val heartrateSensor=childSnapshot.child("heartrateSensor").getValue() as Double
+                val heartrateSensor=childSnapshot.child("heartrateSensor").getValue() as Float
                 val skinTemperatureSensor=childSnapshot.child("skinTemperatureSensor").getValue() as Double
                 if (timestamp != null && heartrateSensor != null && skinTemperatureSensor != null) {
                     val stressData = StressRawData(timestamp, heartrateSensor, skinTemperatureSensor)
                     stressDataList.add(stressData)
                 }
             }
-
             ServiceResult(success = true, data = stressDataList, error = null)
 
-            }
-        catch (e: Exception) {
+        } catch (e: Exception) {
             Log.e("readStressDataFromFirebase", "An exception occurred", e)
             ServiceResult(success = false, data = null, error = StressErrorType.COMMUNICATION_PROBLEM)
         }
-
     }
-    }
+}
